feat(type): add the possibility to set type
This enables setting a type flag on the `oa-sortable` and `oa-sortable-item`. When trying to move an item the items `typeFlag` is validated against the `toSortable.typeFlag` and if the bit is nott set on the sortable the item will not be moved. This enabled having a bucket list and other lists where you could add from the bucket and vice versa but not in between other lists.  You may go crazy :boom: